refactor: 슬랙 앱 설치 관련 예외 클래스 패키지 이동
