+++ backend/src/main/java/com/woowacourse/kkogkkog/exception/infrastructure/BotInstallationFailedException.java
@@ -1,4 +1,4 @@
-package com.woowacourse.kkogkkog.exception.auth;
+package com.woowacourse.kkogkkog.exception.infrastructure;
 
 public class BotInstallationFailedException extends RuntimeException {
