+++ backend/src/test/java/com/woowacourse/kkogkkog/infrastructure/SlackClientTest.java
@@ -6,7 +6,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.woowacourse.kkogkkog.exception.auth.AccessTokenRetrievalFailedException;
-import com.woowacourse.kkogkkog.exception.auth.BotInstallationFailedException;
+import com.woowacourse.kkogkkog.exception.infrastructure.BotInstallationFailedException;
 import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
